refactor: split the docs page into toc, primitives, and prose
Docs/index.tsx held its table of contents, ~360 lines of presentational
primitives, and the documentation prose in one 1279-line file. The TOC
(toc.ts) and the primitives (primitives.tsx) now sit beside it, leaving
index.tsx as the prose it is meant to be (1279 -> 852 lines).

Also trims the db/scans barrel to the types consumers actually import;
knip flagged the rest as unused re-exports.
